Allow selecting from all entries if no match found on autotype shortcut
Fixes #429

If the global autotype shortcut is triggered on a window, but no
applicable match can be found, populate the AutoTypeSelectDialog
with all available entry-sequence pairs. The user can then filter
this list with the filter box implemented in #2955 to narrow down
the matches and select with arrow keys and enter or mouse.